style: rename lib files
