+++ src/lib/api.server.ts
@@ -1,4 +1,4 @@
-import { prismaClient } from '@/lib/prisma';
+import { prismaClient } from '@/lib/prisma.client';
 
 export const getUserLikeCount = async ({
   sessionId,
