+++ src/pages/api/content/[slug].ts
@@ -4,7 +4,7 @@ import { z } from 'zod';
 import { getUserLikeCount } from '@/lib/api.server';
 import { getArticleViewsFromDevto } from '@/lib/devto';
 import { getSessionId } from '@/lib/helper.server';
-import { prismaClient } from '@/lib/prisma';
+import { prismaClient } from '@/lib/prisma.client';
 
 export default async function handler(
   req: NextApiRequest,
